Remove: useless useVATValidation hook
