--- src/hooks/useVATValidation/index.ts
@@ -1,3 +1,0 @@
-import useVATValidation from './useVATValidation'
-
-export default useVATValidation
